refactor: refactoring useVideo
